+++ .gitignore
@@ -3,11 +3,5 @@ dist
 node_modules
 .vscode-test/
 *.vsix
-
-# Ignore pytest cache
 .pytest_cache/
-
-# Ignore Python cache
 __pycache__/
-
-.DS_Store
