+++ README.md
@@ -1,61 +1,71 @@
 # unix-to-utc README
 
-Making time conversion easier and lightweight. The main focus for this extension is for people using Python datetime objects but need the reliability of unix epoch time. I want to make reading unix timestamps easier and lessen time-based project headaches.
+This is the README for your extension "unix-to-utc". After writing up a brief description, we recommend including the following sections.
 
 ## Features
 
-1. Replace unix timestamps with the datetime string version
+Describe specific features of your extension including screenshots of your extension in action. Image paths are relative to this README file.
 
-   - "(Replace) Unix Time with Datetime" Command
+For example if there is an image subfolder under your extension project workspace:
 
-2. Replace datetime timestamps directly (with or without string quotes) as unix timestamps
+\!\[feature X\]\(images/feature-x.png\)
 
-   - "(Replace) Datetime with Unix Time" Command
-
-3. Select a unix timestamp and see a copyable datetime message
-
-   - "(Window) Unix Time to Datetime" Command
-     ![Unix-to-Datetime](images/Screenshot%202023-07-30%20at%208.58.06%20PM.png)
-
-4. Select a Python Datetime string and see a copyable unix message
-
-   - "(Window) Show Datetime to Unix Time
-     ![Datetime-to-Unix](images/Screenshot%202023-07-30%20at%208.58.27%20PM.png)
-
-## Future ideas: 
-   - Terminal Option to convert printed unix times
-   - Flesh out Python Datetime and FastAPI for wider time input support
-   - Select unix time text to see datetime in a hover
+> Tip: Many popular extensions utilize animations. This is an excellent way to show off your extension! We recommend short, focused animations that are easy to follow.
 
 ## Requirements
 
-Goal is to work on any file
+If you have any requirements or dependencies, add a section describing those and how to install and configure them.
 
 ## Extension Settings
 
-Must run commands using cmd(ctrl) + shift + p or and finding the command you want
+Include if your extension adds any VS Code settings through the `contributes.configuration` extension point.
 
-### 1.0.6
+For example:
 
-<!--
+This extension contributes the following settings:
+
+* `myExtension.enable`: Enable/disable this extension.
+* `myExtension.thing`: Set to `blah` to do something.
+
+## Known Issues
+
+Calling out known issues can help limit users opening duplicate issues against your extension.
+
+## Release Notes
+
+Users appreciate release notes as you update your extension.
+
+### 1.0.0
+
+Initial release of ...
+
+### 1.0.1
+
+Fixed issue #.
+
+### 1.1.0
+
+Added features X, Y, and Z.
+
+---
+
 ## Following extension guidelines
 
-Ensure that you've read through the extensions guidelines and follow the best practices for creating your extension. -->
+Ensure that you've read through the extensions guidelines and follow the best practices for creating your extension.
 
-<!--
-- [Extension Guidelines](https://code.visualstudio.com/api/references/extension-guidelines)
+* [Extension Guidelines](https://code.visualstudio.com/api/references/extension-guidelines)
 
 ## Working with Markdown
 
 You can author your README using Visual Studio Code. Here are some useful editor keyboard shortcuts:
 
-- Split the editor (`Cmd+\` on macOS or `Ctrl+\` on Windows and Linux).
-- Toggle preview (`Shift+Cmd+V` on macOS or `Shift+Ctrl+V` on Windows and Linux).
-- Press `Ctrl+Space` (Windows, Linux, macOS) to see a list of Markdown snippets.
+* Split the editor (`Cmd+\` on macOS or `Ctrl+\` on Windows and Linux).
+* Toggle preview (`Shift+Cmd+V` on macOS or `Shift+Ctrl+V` on Windows and Linux).
+* Press `Ctrl+Space` (Windows, Linux, macOS) to see a list of Markdown snippets.
 
 ## For more information
 
-- [Visual Studio Code's Markdown Support](http://code.visualstudio.com/docs/languages/markdown)
-- [Markdown Syntax Reference](https://help.github.com/articles/markdown-basics/)
+* [Visual Studio Code's Markdown Support](http://code.visualstudio.com/docs/languages/markdown)
+* [Markdown Syntax Reference](https://help.github.com/articles/markdown-basics/)
 
-**Enjoy!** -->
+**Enjoy!**
