+++ package.json
@@ -1,63 +1,57 @@
 {
-  "name": "unix-to-datetime-to-epoch",
-  "displayName": "Unix to Datetime to Epoch",
-  "description": "Convert unix time to datetime and datetime to epoch with ease!",
-  "version": "1.1.0",
-  "publisher": "kashsan",
-  "icon": "images/icon.png",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/3sannasia/VSCode-Extension-unix-to-utc.git"
-  },
+  "name": "unix-to-utc",
+  "displayName": "Unix to UTC",
+  "description": "Allows developers to easily read unix timestamps in code by displaying human-readable UTC time inline",
+  "version": "0.0.1",
   "engines": {
-    "vscode": "^1.80.0"
+    "vscode": "^1.85.0"
   },
   "categories": [
     "Other"
   ],
-  "activationEvents": [],
+  "activationEvents": [
+    "onStartupFinished"
+  ],
   "main": "./dist/extension.js",
   "contributes": {
-    "terminal": {
-      "command": "unix-to-datetime-to-epoch.show_datetime",
-      "when": "editorTextFocus"
-    },
     "menus": {
       "editor/context": [
         {
-          "command": "unix-to-datetime-to-epoch.show_datetime",
+          "command": "unix-to-utc.convertTime",
           "when": "editorTextFocus"
         },
         {
-          "command": "unix-to-datetime-to-epoch.show_unix",
+          "command": "unix-to-utc.getUTCTime",
           "when": "editorTextFocus"
         },
         {
-          "command": "unix-to-datetime-to-epoch.replace_unix",
+          "command": "unix-to-utc.getUnixTime",
           "when": "editorTextFocus"
+        }
+      ],
+      "terminal/context": [
+        {
+          "command": "unix-to-utc.getUTCTime",
+          "when": "terminalFocus || terminalIsOpen"
         },
         {
-          "command": "unix-to-datetime-to-epoch.replace_datetime",
-          "when": "editorTextFocus"
+          "command": "unix-to-utc.getUnixTime",
+          "when": "terminalFocus || terminalIsOpen"
         }
       ]
     },
     "commands": [
       {
-        "command": "unix-to-datetime-to-epoch.show_datetime",
-        "title": "(Window) Unix Time to Datetime"
+        "command": "unix-to-utc.convertTime",
+        "title": "Unix <-> UTC"
       },
       {
-        "command": "unix-to-datetime-to-epoch.show_unix",
-        "title": "(Window) Show Datetime to Unix Time"
+        "command": "unix-to-utc.getUTCTime",
+        "title": "Current UTC Time"
       },
       {
-        "command": "unix-to-datetime-to-epoch.replace_unix",
-        "title": "(Replace) Unix Time with Datetime"
-      },
-      {
-        "command": "unix-to-datetime-to-epoch.replace_datetime",
-        "title": "(Replace) Datetime with Unix Time"
+        "command": "unix-to-utc.getUnixTime",
+        "title": "Current Unix Time"
       }
     ]
   },
@@ -70,22 +64,23 @@
     "watch-tests": "tsc -p . -w --outDir out",
     "pretest": "npm run compile-tests && npm run compile && npm run lint",
     "lint": "eslint src --ext ts",
-    "test": "node ./out/test/runTest.js"
+    "test": "vscode-test"
   },
   "devDependencies": {
-    "@types/vscode": "^1.80.0",
-    "@types/glob": "^8.1.0",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "20.2.5",
-    "@typescript-eslint/eslint-plugin": "^5.59.8",
-    "@typescript-eslint/parser": "^5.59.8",
-    "eslint": "^8.41.0",
-    "glob": "^8.1.0",
-    "mocha": "^10.2.0",
-    "typescript": "^5.1.3",
-    "ts-loader": "^9.4.3",
-    "webpack": "^5.85.0",
-    "webpack-cli": "^5.1.1",
-    "@vscode/test-electron": "^2.3.2"
+    "@types/mocha": "^10.0.6",
+    "@types/node": "18.x",
+    "@types/vscode": "^1.85.0",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "^6.15.0",
+    "@vscode/test-cli": "^0.0.4",
+    "@vscode/test-electron": "^2.3.8",
+    "eslint": "^8.56.0",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  },
+  "dependencies": {
+    "node-fetch": "^2.7.0"
   }
 }
