+++ src/extension.ts
@@ -1,12 +1,25 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import * as child_process from "child_process";
+import { exec, ChildProcess } from "node:child_process";
+require("node-fetch");
+import path from "path";
 
-function replaceUnixWithUTC(context: vscode.ExtensionContext): void {
-  const replaceUnix = vscode.commands.registerCommand(
-    "unix-to-datetime-to-epoch.replace_unix",
-    () => {
+let api_process: ChildProcess;
+
+function showCopyWindow(text: string) {
+  vscode.window.showInformationMessage(text, "Copy").then((value) => {
+    if (value === "Copy") {
+      vscode.env.clipboard.writeText(text);
+      vscode.window.showInformationMessage(`Copied ${text}`);
+    }
+  });
+}
+
+async function convertTime(context: vscode.ExtensionContext) {
+  let convertTimeCommand = vscode.commands.registerCommand(
+    "unix-to-utc.convertTime",
+    async () => {
       const editor = vscode.window.activeTextEditor;
 
       if (editor) {
@@ -14,157 +27,86 @@ function replaceUnixWithUTC(context: vscode.ExtensionContext): void {
         const selection = editor.selection;
 
         const text = document.getText(selection);
-        const unix = parseFloat(text);
-
-        if (
-          isNaN(unix) ||
-          unix < 0 ||
-          selection.isEmpty ||
-          !selection.isSingleLine
-        ) {
-          vscode.window.showInformationMessage(
-            "Please select a valid unix timestamp!"
-          );
-          return;
+        console.log(text);
+        console.log("http://127.0.0.1:8001/convert/" + text.toString()) + "/";
+        const response = await fetch(
+          `http://127.0.0.1:8001/convert/${text.toString()}`
+        );
+        if (!response.ok) {
+          throw new Error(`HTTP error! Status: ${response.status}`);
         }
-
-        const datetime = UnixToDatetime(unix);
-        editor.edit((editBuilder) => {
-          editBuilder.replace(selection, datetime);
-        });
+        const data: any = await response.json();
+        const converted_time = data["date"];
+        showCopyWindow(converted_time.toString());
+        // editor.edit((editBuilder) => {
+        //   editBuilder.replace(selection, converted_time.toString());
+        // });
       }
     }
   );
-
-  context.subscriptions.push(replace_unix);
+  context.subscriptions.push(convertTimeCommand);
 }
 
-function showDatetimeFromUnix(context: vscode.ExtensionContext): void {
-  const convert = vscode.commands.registerCommand(
-    "unix-to-utc.show_datetime",
-    () => {
-      const editor = vscode.window.activeTextEditor;
-
-      if (editor) {
-        const document = editor.document;
-        const selection = editor.selection;
-
-        const text = document.getText(selection);
-        const unix_float = parseFloat(text);
-
-        if (
-          isNaN(unix_float) ||
-          unix_float < 0 ||
-          selection.isEmpty ||
-          !selection.isSingleLine
-        ) {
-          vscode.window.showInformationMessage(
-            "Please select a valid unix timestamp!"
-          );
-          return;
-        }
-
-        const datetime = UnixToDatetime(unix_float);
-
-        vscode.window.showInformationMessage(datetime, "Copy").then((value) => {
-          if (value === "Copy") {
-            vscode.env.clipboard.writeText(datetime);
-            vscode.window.showInformationMessage(`Copied ${datetime}`);
-          }
-        });
+async function getUTCTime(context: vscode.ExtensionContext) {
+  let getUTCTimeCommand = vscode.commands.registerCommand(
+    "unix-to-utc.getUTCTime",
+    async () => {
+      const response = await fetch(
+        "http://127.0.0.1:8001/current_utc_timestamp"
+      );
+      if (!response.ok) {
+        throw new Error(`HTTP error! Status: ${response.status}`);
       }
+      const data: any = await response.json();
+      const cur_utc = data["date"];
+      showCopyWindow(cur_utc.toString());
     }
   );
-  context.subscriptions.push(convert);
+
+  context.subscriptions.push(getUTCTimeCommand);
 }
 
-function replaceDatetimeWithUnix(context: vscode.ExtensionContext) {
-  const replace_datetime = vscode.commands.registerCommand(
-    "unix-to-datetime-to-epoch.replace_datetime",
-    () => {
-      const editor = vscode.window.activeTextEditor;
-
-      if (editor) {
-        const document = editor.document;
-        const selection = editor.selection;
-
-        var text = document.getText(selection);
-
-        if (text.at(0) == '"' && text.at(text.length - 1) == '"') {
-          text = text.slice(1, text.length - 1);
-        }
-
-        const date = DatetimeToDate(text);
-        const unix = DateToUnix(date);
-
-        if (isNaN(unix)) {
-          vscode.window.showInformationMessage(
-            "Please select a valid datetime!"
-          );
-          return;
-        }
-
-        editor.edit((editBuilder) => {
-          editBuilder.replace(selection, unix.toString());
-        });
+async function getUnixTime(context: vscode.ExtensionContext) {
+  let getUnixTimeCommand = vscode.commands.registerCommand(
+    "unix-to-utc.getUnixTime",
+    async () => {
+      const response = await fetch(
+        "http://127.0.0.1:8001/current_unix_timestamp"
+      );
+      if (!response.ok) {
+        throw new Error(`HTTP error! Status: ${response.status}`);
       }
+      const data: any = await response.json();
+      const cur_unix = data["date"].toString();
+      showCopyWindow(cur_unix.toString());
     }
   );
-  context.subscriptions.push(replace_datetime);
+  context.subscriptions.push(getUnixTimeCommand);
 }
 
-// function showUnixFromUTC(context: vscode.ExtensionContext) {
-//   const convert = vscode.commands.registerCommand(
-//     "unix-to-datetime-to-epoch.show_unix",
-//     () => {
-//       const editor = vscode.window.activeTextEditor;
-
-//       if (editor) {
-//         const document = editor.document;
-//         const selection = editor.selection;
-
-//         var text = document.getText(selection);
-
-//         if (text.at(0) == '"' && text.at(text.length - 1) == '"') {
-//           text = text.slice(1, text.length - 1);
-//         }
-
-//         const date = DatetimeToDate(text);
-//         const unix = DateToUnix(date);
-
-//         if (isNaN(unix)) {
-//           vscode.window.showInformationMessage(
-//             "Please select a valid datetime!"
-//           );
-//           return;
-//         }
-//         vscode.window
-//           .showInformationMessage(unix.toString(), "Copy")
-//           .then((value) => {
-//             if (value === "Copy") {
-//               vscode.env.clipboard.writeText(unix.toString());
-//               vscode.window.showInformationMessage(`Copied ${unix}`);
-//             }
-//           });
-//       }
-//     }
-//   );
-//   context.subscriptions.push(convert);
-// }
-
-// Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-  vscode.window.showInformationMessage("Unix to UTC activated");
+  vscode.window.showInformationMessage("Unix to UTC is now active");
+  api_process = exec(
+    "python3 " + path.join(__dirname, "datetime_api.py"),
+    (error, stderr) => {
+      if (error) {
+        console.error(`Error executing script: ${error.message}`);
+        return;
+      }
 
-  // Commands to convert between Unix time and Python datetime objects
-  showUnixFromUTC(context);
-  showDatetimeFromUnix(context);
-
-  replaceUnixWithUTC(context);
-  replaceDatetimeWithUnix(context);
+      if (stderr) {
+        console.error(`Script execution resulted in an error: ${stderr}`);
+        return;
+      }
+    }
+  );
+  console.log("API started at http://127.0.0.1:8001");
+  convertTime(context);
+  getUTCTime(context);
+  getUnixTime(context);
 }
 
-// This method is called when your extension is deactivated
 export function deactivate() {
-  vscode.window.showInformationMessage("Deactivated unix-to-utc");
+  console.log("Deactivated extension. Killing API process.");
+  api_process.kill("SIGINT");
 }
