+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
 	"compilerOptions": {
-		"module": "commonjs",
-		"target": "ES2020",
+		"module": "Node16",
+		"target": "ES2022",
 		"lib": [
-			"ES2020"
+			"ES2022"
 		],
 		"sourceMap": true,
 		"rootDir": "src",
