--- unix-to-utc/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "unix-to-utc",
-  "displayName": "Unix to UTC",
-  "description": "Allows developers to easily read unix timestamps in code by displaying human-readable UTC time inline",
-  "version": "0.0.1",
-  "engines": {
-    "vscode": "^1.85.0"
-  },
-  "categories": [
-    "Other"
-  ],
-  "activationEvents": [
-    "onStartupFinished"
-  ],
-  "main": "./dist/extension.js",
-  "contributes": {
-    "menus": {
-      "editor/context": [
-        {
-          "command": "unix-to-utc.getUTCTime",
-          "when": "editorTextFocus"
-        },
-        {
-          "command": "unix-to-utc.getUnixTime",
-          "when": "editorTextFocus"
-        },
-        {
-          "command": "unix-to-utc.convertTime",
-          "when": "editorTextFocus"
-        }
-      ]
-    },
-    "commands": [
-      {
-        "command": "unix-to-utc.convertTime",
-        "title": "Unix <-> UTC"
-      },
-      {
-        "command": "unix-to-utc.getUTCTime",
-        "title": "Current UTC Time"
-      },
-      {
-        "command": "unix-to-utc.getUnixTime",
-        "title": "Current Unix Time"
-      }
-    ]
-  },
-  "scripts": {
-    "vscode:prepublish": "npm run package",
-    "compile": "webpack",
-    "watch": "webpack --watch",
-    "package": "webpack --mode production --devtool hidden-source-map",
-    "compile-tests": "tsc -p . --outDir out",
-    "watch-tests": "tsc -p . -w --outDir out",
-    "pretest": "npm run compile-tests && npm run compile && npm run lint",
-    "lint": "eslint src --ext ts",
-    "test": "vscode-test"
-  },
-  "devDependencies": {
-    "@types/mocha": "^10.0.6",
-    "@types/node": "18.x",
-    "@types/vscode": "^1.85.0",
-    "@typescript-eslint/eslint-plugin": "^6.15.0",
-    "@typescript-eslint/parser": "^6.15.0",
-    "@vscode/test-cli": "^0.0.4",
-    "@vscode/test-electron": "^2.3.8",
-    "eslint": "^8.56.0",
-    "ts-loader": "^9.5.1",
-    "typescript": "^5.3.3",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4"
-  },
-  "dependencies": {
-    "node-fetch": "^2.7.0"
-  }
-}
