--- unix-to-utc/src/api.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-python3 ../dist/datetime_api.py
