--- unix-to-utc/src/test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-pytest test_datetime_api.py
